refactor: remove legacy components and implement new UI
