+++ webapp/src/webapp/db.cljs
@@ -119,8 +119,4 @@
    :provisioning {:resources {:status :idle :data []}
                   :jobs      []
                   :sessions  []}
-   :rulepacks {:list {:status :idle :data []}
-               :active {:status :idle :data nil}
-               :selected-connections #{}
-               :applying? false}
    :connections->list {:status :idle :data []}})
